Integrating Igor's and Francisco's comments.
