+++ tso-server/src/main/java/com/yahoo/omid/tso/TSOModule.java
@@ -16,12 +16,8 @@
 package com.yahoo.omid.tso;
 
 import com.google.inject.AbstractModule;
+import com.google.inject.Provider;
 import com.google.inject.Provides;
-import com.google.inject.name.Names;
-import com.yahoo.omid.committable.CommitTable;
-import com.yahoo.omid.committable.hbase.HBaseCommitTableConfig;
-import com.yahoo.omid.timestamp.storage.TimestampStorage;
-import com.yahoo.omid.tools.hbase.SecureHBaseConfig;
 
 import javax.inject.Named;
 import javax.inject.Singleton;
@@ -63,4 +59,12 @@ class TSOModule extends AbstractModule {
 
     }
 
+    @Provides
+    PersistenceProcessorHandler[] getPersistenceProcessorHandler(Provider<PersistenceProcessorHandler> provider) {
+        PersistenceProcessorHandler[] persistenceProcessorHandlers = new PersistenceProcessorHandler[config.getPersistHandlerNum()];
+        for (int i = 0; i < persistenceProcessorHandlers.length; i++) {
+            persistenceProcessorHandlers[i] = provider.get();
+        }
+        return persistenceProcessorHandlers;
+    }
 }
